GUI Calendar using Tkinter
Prerequisites: Introduction to Tkinter

Python offers multiple options for developing a GUI (Graphical User Interface). Out of all the GUI methods, Tkinter is the most commonly used method. Python with Tkinter outputs the fastest and easiest way to create GUI applications. In this article, we will learn how to create a GUI Calendar application using Tkinter, with a step-by-step guide. 

To create a Tkinter: 

Importing the module – Tkinter
Create the main window (container)
Add any number of widgets to the main window.
Apply the event Trigger on the widgets.