Bubble sort visualizer using PyGame
hen the pygame application get started we can see the unsorted bars with different heights and when we click space bar key it started getting arranging in bubble sort manner i.e after every iteration maximum value element should come at last. Bubble Sort is a simple algorithm which is used to sort a given set of n elements provided in form of an array with n number of elements. Bubble Sort compares all the element one by one and sort them based on their values.

Implementation steps:
Create a main window
Fill the main window with black color
Create a method to show the list of bar with specific gap in between them
Get the keys input from the user 5. If space bar is pressed start the sorting process
Implement bubble sort algorithm on the list
After every internal iteration fill the screen with black color and call the show method to show the iterated list in the form of bar.