Screen recorder using Python
Python is a widely used general-purpose language. It allows for performing a variety of tasks. One of them can be recording a video. It provides a module named pyautogui which can be used for the same. This module along with NumPy and OpenCV provides a way to manipulate and save the images (screenshot in this case)

Modules needed
Numpy: To install Numpy type the below command in the terminal.
pip install numpy
pyautogui: To install pyautogui type the below command in the terminal.
pip install pyautogui
OpenCV: To install OpenCV type the below command in the terminal.
pip install opencv-python